ICC92_classifier {ICC92clusfier}	R Documentation

Dynamic knapsack.

Description

Creating dynamic algorithm to solve knapsack problem.

Usage

ICC92_classifier(validate, eta = 0.1, nrounds = 3)

Arguments

validate
expression matrix

eta
pass to xgboost/ default = 0.1

nrounds
pass to xgboost. default=3

Details

Input takes a data.frame x with two variables v and w and returns the maximum knapsack value and which elements (rows in the data.frame).

Value

A numeric vector

[Package ICC92clusfier version 0.1.0 Index]